feat: ATL-151: unify tests with existing ones
